feat(Gallery): add handle click to bullet
