+++ packages/vkui/src/components/BaseGallery/BaseGallery.module.css
@@ -74,6 +74,10 @@
   margin-block: 0;
   margin-inline: 3px;
   opacity: var(--vkui--opacity_disable_accessibility);
+  border: 0;
+  outline: 0;
+  padding: 0;
+  cursor: pointer;
 }
 
 .BaseGallery__bullet--active {
